Fixed an issue where my int64s were not int64_ts

+++ ps05/cache/cache.c
@@ -8,14 +8,15 @@ void main() {
    @param cache_size The expected size of the cache
    @return the array representation of the timing of our cache
 */
-double* testacache(int64 cache_size) {
-  int64 array_size = (cache_size / 8) + 1;
-  int64 arrayOfTestValues[array_size];
+double* testacache(int64_t cache_size) {
+  int64_t array_size = (cache_size / 8) + 1;
+  int64_t arrayOfTestValues[array_size];
   for (int step = 0; step < array_size; step++) {
-    arrayOfTestValues[step] = step;
+    arrayOfTestValues[step] = (int64_t)step;
   }
   /* Warm up the cache by moving through the array once */
-  int64 x = random_int64();
+  /* We need ot write a function to create a random int64 */
+  int64_t x = random_int64();
   for (int i = 0; i < 10*(array_size); i++) {
     x = x ^ arrayOfTestValues[i];
   }
